scripts/breaking-changes/requirements.txt
langchain
langchain-google-genai
requests
beautifulsoup4
python-dotenv
